45899877
git-svn-id: svn+ssh://lo-svn/opt/localorbit/lo3/trunk@1263 e1dfb558-8c25-43e7-bb2c-da0e82dfb25a

+++ www/app/controllers/payments/views/receivables__enter_invoices.php
@@ -122,8 +122,8 @@ foreach($invoices as $invoice)
 					<?=$invoice['description_html']?>
 				</td>
 				<td class="dt"><?=$invoice['direction_info']?></td>
-				<td class="dt"><?=core_format::date($invoice['creation_date'],'long')?></td></td>
-				<td class="dt"><?=core_format::date($invoice['due_date'],'long')?></td></td>
+				<td class="dt"><?=core_format::date($invoice['creation_date'],'short')?></td></td>
+				<td class="dt"><?=core_format::date($invoice['due_date'],'short')?></td></td>
 				<td class="dt"><?=$invoice['age']?></td></td>
 				<td class="dt">
 					<input type="text" class="pull-left" name="<?=$core->data['tab_name']?>_invoice_<?=$invoice['invoice_id']?>" style="width: 120px;" />
